[FEAT] Read parquet tables with int96 coercion option (#1231)
Add option to coerce int96 timestamps to a specific TimeUnit during
reads of Parquet

---------

+++ tests/dataframe/test_creation.py
@@ -760,10 +760,10 @@ def test_create_dataframe_parquet_specify_schema(
         df = daft.read_parquet(
             f.name,
             schema_hints={
-                "sepal_length": DataType.float32(),
-                "sepal_width": DataType.float32(),
-                "petal_length": DataType.float32(),
-                "petal_width": DataType.float32(),
+                "sepal_length": DataType.float64(),
+                "sepal_width": DataType.float64(),
+                "petal_length": DataType.float64(),
+                "petal_width": DataType.float64(),
                 "variety": DataType.string(),
             },
             use_native_downloader=use_native_downloader,
